+++ src/functions/parseKV.ts
@@ -1,7 +1,7 @@
 import ValueCollection from "@/classes/ValueCollection"
 import { as } from "@rjweb/utils"
 
-export const trimString = (str: string): string => {
+export function trimString(str: string): string {
   let start = 0, end = str.length - 1
 
   while (start < end && str[start] === ' ') {